FIX properly register memcached instances
When memcached servers are started from environment configuration, they
did not register themselves to the pool, and thus pool appeared to not
exist.

Fix that by introducing the global pool registration table and
registering the children by themselves.

"global registry table" might sound inelegant, but it is dual to having
a dedicated process for that registration, but with less code. If, in
the future, we have more consumers of the table, it might make sense to
make it a process.